Re #115: Attempt to fix linux build error
git-svn-id: https://subversion.assembla.com/svn/slicerrt/trunk/SlicerRt/src@2694 25a7314f-c4c8-4314-9936-a25f8480a772

+++ SlicerRtCommon/vtkCollisionDetectionFilter.cxx
@@ -521,7 +521,7 @@ int vtkCollisionDetectionFilter::RequestData(
   tmpMatrix->Delete();
 
   vtkDebugMacro(<< "Collision detection finished");
-  this->NumberOfBoxTests = std::abs(BoxTests);
+  this->NumberOfBoxTests = std::abs((int)BoxTests);
   
   // Generate the scalars if needed
   if (GenerateScalars)
